0-print_list.c

+++ 0x12-singly_linked_lists/lists.h
@@ -13,7 +13,7 @@
 typedef struct list
 {
 	char *name;
-	size_t len;
+	unsigned int len;
 	struct list *next;
 }list_t;
